requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
nltk
summa
pymupdf==1.22.5  # Specify the latest stable version
pillow
python-docx
requests
beautifulsoup4
openpyxl
